fix: resolves RulesObject type and circular dependency
