+++ packages/flower-core/src/devtoolState.ts
@@ -1,0 +1,2 @@
+const devtoolState = {}
+export { devtoolState }
