+++ packages/flower-core/src/index.ts
@@ -4,7 +4,5 @@ export { FlowerStateUtils } from './FlowerCoreStateUtils'
 export { FlowerCoreStateSelectors as Selectors } from './FlowerCoreStateSelectors'
 export { CoreUtils } from './CoreUtils'
 export { MatchRules } from './RulesMatcher'
-
+export { devtoolState } from './devtoolState'
 export * from './interfaces'
-const devtoolState = {}
-export { devtoolState }
